refactor: posts.ts 파일에서 타입 분리 및 PostListType.ts 객체 타입 interface로 수정
